fixed data types

+++ marketplace/src/main/java/com/ecommerce/marketplace/dto/product/productListingDTO.java
@@ -1,16 +1,12 @@
 package com.ecommerce.marketplace.dto.product;
 
 
-import com.ecommerce.marketplace.entities.product.products;
-import com.ecommerce.marketplace.entities.seller.seller;
-import jakarta.persistence.Column;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -27,7 +23,7 @@ public class productListingDTO {
     private String seller;
 
     @Min(value = 1, message = "pricing  should least be 1")
-    private Double price;
+    private BigDecimal price;
 
     @Min(value = 1,message = "stock quantity is mandatory")
     private Integer stockQuantity;
